Remove IsoStorage lock from UAP and enable tests
Removing the random directory logic when running on WinRT as
it wasn't used before and has no benefit. Additionally taking
a global lock isn't possible on WinRT.

This also enables tests for IsolatedStorage on UAP. Also update
skip for unit tests on desktop.

Issues #19227, #18202
