+++ src/System.IO.IsolatedStorage/src/System.IO.IsolatedStorage.csproj
@@ -29,7 +29,7 @@
     <Compile Include="System\IO\IsolatedStorage\Helper.Win32Unix.cs" />
   </ItemGroup>
   <ItemGroup Condition="'$(TargetGroup)' != 'uap' AND '$(TargetsWindows)' == 'true'">
-	<Compile Include="System\IO\IsolatedStorage\Helper.Win32.cs" />
+    <Compile Include="System\IO\IsolatedStorage\Helper.Win32.cs" />
   </ItemGroup>
   <ItemGroup Condition="'$(TargetGroup)' != 'uap' AND '$(TargetsUnix)' == 'true'">
     <Compile Include="System\IO\IsolatedStorage\Helper.Unix.cs" />
