+++ src/System.IO.IsolatedStorage/src/System/IO/IsolatedStorage/Helper.cs
@@ -2,9 +2,6 @@
 // The .NET Foundation licenses this file to you under the MIT license.
 // See the LICENSE file in the project root for more information.
 
-using System.Security;
-using System.Threading;
-
 namespace System.IO.IsolatedStorage
 {
     internal static partial class Helper
@@ -17,7 +14,8 @@ namespace System.IO.IsolatedStorage
 
         /// <summary>
         /// The full root directory is the relevant special folder from Environment.GetFolderPath() plus "IsolatedStorage"
-        /// and a set of random directory names if not roaming.
+        /// and a set of random directory names if not roaming. (The random directories aren't created for WinRT as
+        /// the FolderPath locations for WinRT are app isolated already.)
         /// 
         /// Examples:
         /// 
@@ -51,71 +49,6 @@ namespace System.IO.IsolatedStorage
                 s_userRootDirectory = GetRandomDirectory(GetDataDirectory(scope), scope);
 
             return s_userRootDirectory;
-        }
-
-        internal static string GetRandomDirectory(string rootDirectory, IsolatedStorageScope scope)
-        {
-            string randomDirectory = GetExistingRandomDirectory(rootDirectory);
-            if (string.IsNullOrEmpty(randomDirectory))
-            {
-                using (Mutex m = CreateMutexNotOwned(rootDirectory))
-                {
-                    if (!m.WaitOne())
-                    {
-                        throw new IsolatedStorageException(SR.IsolatedStorage_Init);
-                    }
-
-                    try
-                    {
-                        randomDirectory = GetExistingRandomDirectory(rootDirectory);
-                        if (string.IsNullOrEmpty(randomDirectory))
-                        {
-                            // Someone else hasn't created the directory before we took the lock
-                            randomDirectory = Path.Combine(rootDirectory, Path.GetRandomFileName(), Path.GetRandomFileName());
-                            CreateDirectory(randomDirectory, scope);
-                        }
-                    }
-                    finally
-                    {
-                        m.ReleaseMutex();
-                    }
-                }
-            }
-
-            return randomDirectory;
-        }
-
-        internal static string GetExistingRandomDirectory(string rootDirectory)
-        {
-            // Look for an existing random directory at the given root
-            // (a set of nested directories that were created via Path.GetRandomFileName())
-
-            // Older versions of the desktop framework created longer (24 character) random paths and would
-            // migrate them if they could not find the new style directory.
-
-            if (!Directory.Exists(rootDirectory))
-                return null;
-
-            foreach (string directory in Directory.GetDirectories(rootDirectory))
-            {
-                if (Path.GetFileName(directory)?.Length == 12)
-                {
-                    foreach (string subdirectory in Directory.GetDirectories(directory))
-                    {
-                        if (Path.GetFileName(subdirectory)?.Length == 12)
-                        {
-                            return subdirectory;
-                        }
-                    }
-                }
-            }
-
-            return null;
-        }
-
-        private static Mutex CreateMutexNotOwned(string pathName)
-        {
-            return new Mutex(initiallyOwned: false, name: @"Global\" + IdentityHelper.GetStrongHashSuitableForObjectName(pathName));
         }
 
         internal static bool IsMachine(IsolatedStorageScope scope) => ((scope & IsolatedStorageScope.Machine) != 0);
