+++ src/System.IO.IsolatedStorage/tests/Configurations.props
@@ -6,6 +6,7 @@
       netstandard-Windows_NT;
       netcoreapp-Unix;
       netcoreapp-Windows_NT;
+      uap-Windows_NT;
     </BuildConfigurations>
   </PropertyGroup>
 </Project>
